+++ packages/babel-plugin-proposal-private-property-in-object/test/fixtures/assumption-privateFieldsAsProperties/static-accessor/output.js
@@ -1,7 +1,7 @@
 var _foo = /*#__PURE__*/babelHelpers.classPrivateFieldLooseKey("foo");
 class Foo {
   test(other) {
-    return Object.prototype.hasOwnProperty.call(other, _foo);
+    return Object.prototype.hasOwnProperty.call(babelHelpers.checkInRHS(other), _foo);
   }
 }
 function _get_foo() {}