+++ packages/babel-plugin-proposal-private-property-in-object/test/fixtures/private/nested-class-redeclared/output.js
@@ -24,12 +24,12 @@ let Foo = /*#__PURE__*/function () {
         babelHelpers.createClass(Nested, [{
           key: "test",
           value: function test() {
-            _foo2.has(this);
+            _foo2.has(babelHelpers.checkInRHS(this));
           }
         }]);
         return Nested;
       }();
-      _foo.has(this);
+      _foo.has(babelHelpers.checkInRHS(this));
     }
   }]);
   return Foo;