+++ packages/babel-plugin-proposal-private-property-in-object/test/fixtures/private/static-method/output.js
@@ -7,7 +7,7 @@ let Foo = /*#__PURE__*/function () {
   babelHelpers.createClass(Foo, [{
     key: "test",
     value: function test(other) {
-      return other === Foo;
+      return babelHelpers.checkInRHS(other) === Foo;
     }
   }]);
   return Foo;