+++ packages/babel-plugin-transform-block-scoping/test/fixtures/general/switch-shadow/input.js
@@ -1,0 +1,3 @@
+switch (x) {
+  default: let x;
+}