+++ packages/babel-plugin-transform-block-scoping/test/fixtures/tdz/switch-shadow/output.js
@@ -1,0 +1,5 @@
+var x;
+switch (x) {
+  default:
+    var _x;
+}