+++ packages/babel-plugin-transform-classes/test/fixtures/get-set-loose/call-semantics-not-defined-on-parent/input.js
@@ -18,4 +18,4 @@ expect(() => {
 
   // Asser that this throws, but that it's not
   // Obj.p.test's error that is thrown
-}).toThrowError(TypeError)
+}).toThrow(TypeError)