+++ packages/babel-plugin-transform-classes/test/fixtures/get-set/call-semantics-setter-defined-on-parent/output.js
@@ -38,4 +38,4 @@ expect(() => {
 
   // Assert that this throws, but that it's not
   // a gobbledygook error that is thrown
-}).toThrowError(TypeError);
+}).toThrow(TypeError);