+++ packages/babel-plugin-transform-regenerator/test/fixtures/regression/15142/exec.js
@@ -1,0 +1,4 @@
+var f = async () => [await 0, , ];
+f().then(result => {
+  expect(result).toStrictEqual([0, , ]);
+});