+++ packages/babel-plugin-transform-regenerator/test/fixtures/regression/15177/input.js
@@ -1,0 +1,7 @@
+var [x] = {
+  [Symbol.iterator]: function* () {
+    yield* {
+      [Symbol.iterator]: x => [],
+    };
+  },
+};